Rename MasterMock header only file
