+++ src/tests/CMakeLists.txt
@@ -23,7 +23,7 @@ $<TARGET_OBJECTS:kafka_to_nexus__objects>
 )
 set(UnitTests_INC
 HDFFileTestHelper.h
-MM.h
+MasterMock.h
 )
 
 add_executable(UnitTests EXCLUDE_FROM_ALL ${UnitTests_SRC} ${UnitTests_INC})
